Refactor: Table 컴포넌트 정리
