+++ src/components/Table/Table.styles.ts
@@ -1,16 +1,7 @@
 import styled, { DefaultTheme } from 'styled-components';
-import { CiSearch } from 'react-icons/ci';
 
 export const TableContainer = styled.div`
   width: 100%;
-`;
-
-export const TableNav = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-
-  margin: 1.625rem 0;
 `;
 
 export const TableBox = styled.div`
@@ -18,73 +9,6 @@ export const TableBox = styled.div`
   background-color: ${({ theme }) => theme.colors.bg_element4};
   border-radius: 15px;
   box-shadow: 0px 0px 25px 0px rgba(48, 73, 191, 0.07);
-`;
-
-export const SelectWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 0.9375rem;
-
-  color: ${({ theme }) => theme.colors.text1};
-  font-family: Uni Neue-Trial;
-  font-size: 0.875rem;
-  font-style: normal;
-  font-weight: 700;
-  line-height: normal;
-`;
-
-export const SearchWrapper = styled.div`
-  display: flex;
-  align-items: center;
-`;
-
-export const CoinCount = styled.div`
-  color: ${({ theme }) => theme.colors.text1};
-  font-family: Uni Neue-Trial;
-  font-size: 0.875rem;
-  font-style: normal;
-  font-weight: 700;
-  line-height: normal;
-
-  margin-right: 1.25rem;
-`;
-
-export const SearchForm = styled.form`
-  display: flex;
-  align-items: center;
-  padding: 0.5625rem 1rem;
-
-  width: 232px;
-  height: 36px;
-  border-radius: 10px;
-  border: 1px solid ${({ theme }) => theme.colors.border1};
-  background: ${({ theme }) => theme.colors.bg_element4};
-  box-shadow: 0px 0px 40px 0px ${({ theme }) => theme.colors.border1};
-
-  &:focus-within {
-    border: 1px solid ${({ theme }) => theme.colors.blue};
-  }
-`;
-
-export const SearchIcon = styled(CiSearch)`
-  width: 18px;
-  height: 18px;
-  margin-right: 0.625rem;
-`;
-
-export const SearchInput = styled.input`
-  width: 100%;
-  border: none;
-  outline: none;
-
-  color: ${({ theme }) => theme.colors.text1};
-  font-family: Pretendard;
-  font-size: 0.75rem;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-
-  background-color: inherit;
 `;
 
 export const CoinListWrapper = styled.div`
