+++ src/components/Table/TableNav/index.ts
@@ -1,0 +1,1 @@
+export { TableNav } from './TableNav';
